refactor(custom-hook): refactor useBlockGroups
- 데이터를 전역상태에서 받아오는 쪽과, 조작하는 쪽을 분리하는 작업을 수행했다.
